* More flexible handlers * Handlers made flexible * Central register * Inherited base module * Remove '!!python/unicode' from output of name when mass converting
